Moved websocket CreateXXX to websocketHelpers.js
