+++ src/index.js
@@ -22,6 +22,7 @@ const parseResources = require('./parseResources');
 const { createDefaultApiKey, detectEncoding, randomId } = require('./utils');
 const authFunctionNameExtractor = require('./authFunctionNameExtractor');
 const requestBodyValidator = require('./requestBodyValidator');
+const wsHelpers = require('./websocketHelpers');
 
 /*
   I'm against monolithic code like this file
@@ -472,111 +473,6 @@ class Offline {
         });
       }
     };
-
-    const createRequestContext = (action, eventType, connection) => {
-      const now = new Date();
-      const months = ['January', 'February', 'March', 'April', 'May', 'June', 'July', 'August', 'September', 'October', 'November', 'December'];
-      const requestContext = { 
-        routeKey: action,
-        messageId: `${randomId()}`,
-        eventType,
-        extendedRequestId: `${randomId()}`,
-        requestTime: `${now.getUTCDate()}/${months[now.getUTCMonth()]}/${now.getUTCFullYear()}:${now.getUTCHours()}:${now.getUTCMinutes()}:${now.getSeconds()} +0000`,
-        messageDirection: 'IN',
-        stage: 'local',
-        connectedAt: connection.connectionTime,
-        requestTimeEpoch: now.getTime(),
-        identity:
-          { cognitoIdentityPoolId: null,
-            cognitoIdentityId: null,
-            principalOrgId: null,
-            cognitoAuthenticationType: null,
-            userArn: null,
-            userAgent: null,
-            accountId: null,
-            caller: null,
-            sourceIp: '127.0.0.1',
-            accessKey: null,
-            cognitoAuthenticationProvider: null,
-            user: null },
-        requestId: `${randomId()}`,
-        domainName: 'localhost',
-        connectionId:connection.connectionId,
-        apiId: 'private', 
-      };
-      
-      return requestContext;
-    };
-
-    const createEvent = (action, eventType, connection, payload) => {
-      const event = { 
-        requestContext: createRequestContext(action, eventType, connection),
-        body: JSON.stringify(payload),
-        isBase64Encoded: false,
-        apiGatewayUrl: `http${this.options.httpsProtocol ? 's' : ''}://${this.options.host}:${this.options.port + 1}`,
-      };
-      
-      return event;
-    };
-
-    const createConnectEvent = (action, eventType, connection) => {
-      const headers = { 
-        Host: 'localhost',
-        'Sec-WebSocket-Extensions': 'permessage-deflate; client_max_window_bits',
-        'Sec-WebSocket-Key': `${randomId()}`,
-        'Sec-WebSocket-Version': '13',
-        'X-Amzn-Trace-Id': `Root=${randomId()}`,
-        'X-Forwarded-For': '127.0.0.1',
-        'X-Forwarded-Port': `${this.options.port + 1}`,
-        'X-Forwarded-Proto': `http${this.options.httpsProtocol ? 's' : ''}`, 
-      };
-      const multiValueHeaders = { ...headers };
-      Object.keys(multiValueHeaders).map(key => multiValueHeaders[key] = [multiValueHeaders[key]]);
-      const event = { 
-        headers,
-        multiValueHeaders,
-        requestContext: createRequestContext(action, eventType, connection),
-        apiGatewayUrl: `http${this.options.httpsProtocol ? 's' : ''}://${this.options.host}:${this.options.port + 1}`,
-        isBase64Encoded: false,
-      };
-        
-      return event;
-    };
-
-    const createDisconnectEvent = (action, eventType, connection) => {
-      const headers = { 
-        Host: 'localhost',
-        'x-api-key': '',
-        'x-restapi': '',
-      };
-      const multiValueHeaders = { ...headers };
-      Object.keys(multiValueHeaders).map(key => multiValueHeaders[key] = [multiValueHeaders[key]]);
-      const event = { 
-        headers,
-        multiValueHeaders,
-        requestContext: createRequestContext(action, eventType, connection),
-        apiGatewayUrl: `http${this.options.httpsProtocol ? 's' : ''}://${this.options.host}:${this.options.port + 1}`,
-        isBase64Encoded: false,
-      };
-        
-      return event;
-    };
-
-    const createContext = action => {
-      const context = {
-        awsRequestId: `offline_awsRequestId_for_${action}`,
-        callbackWaitsForEmptyEventLoop: true,
-        functionName: action,
-        functionVersion: '$LATEST',
-        invokedFunctionArn: `offline_invokedFunctionArn_for_${action}`,
-        invokeid: `offline_invokeid_for_${action}`,
-        logGroupName: `offline_logGroupName_for_${action}`,
-        logStreamName: `offline_logStreamName_for_${action}`,
-        memoryLimitInMB: '1024',
-      };
-
-      return context;
-    };
     
     this.wsServer.route({
       method: 'POST', 
@@ -605,9 +501,9 @@ class Offline {
               debugLog(`connect:${connection.connectionId}`);
 
               this.clients.set(ws, connection);
-              let event = createConnectEvent('$connect', 'CONNECT', connection);
+              let event = wsHelpers.createConnectEvent('$connect', 'CONNECT', connection, this.options);
               if (Object.keys(queryStringParameters).length > 0) event = { queryStringParameters, ...event };
-              const context = createContext('$connect');
+              const context = wsHelpers.createContext('$connect');
 
               doAction(ws, connection.connectionId, '$connect', event, context);
             },
@@ -615,8 +511,8 @@ class Offline {
               const connection = this.clients.get(ws);
               debugLog(`disconnect:${connection.connectionId}`);
               this.clients.delete(ws);
-              const event = createDisconnectEvent('$disconnect', 'DISCONNECT', connection);
-              const context = createContext('$disconnect');
+              const event = wsHelpers.createDisconnectEvent('$disconnect', 'DISCONNECT', connection, this.options);
+              const context = wsHelpers.createContext('$disconnect', this.options);
 
               doAction(ws, connection.connectionId, '$disconnect', event, context);
             },
@@ -629,8 +525,8 @@ class Offline {
         const connection = this.clients.get(ws);
         const action = request.payload.action || '$default';
         debugLog(`action:${action} on connection=${connection.connectionId}`);
-        const event = createEvent(action, 'MESSAGE', connection, request.payload);
-        const context = createContext(action);
+        const event = wsHelpers.createEvent(action, 'MESSAGE', connection, request.payload, this.options);
+        const context = wsHelpers.createContext(action, this.options);
 
         doAction(ws, connection.connectionId, action, event, context, true);
